Improve search for drkonqui and keep it silent per default if not found
Search not only in env var + install prefix but in application binary directory and libexec path as configured for qt (or qt.conf).
That allows it to bundle drkonqui if wanted.
Beside, make the output of "not found" an per default hidden debug message, as otherwise, any application using kcrash without workspace stuff installed warns on start.
Sideeffect: decodeName used to decode the CMAKE var, like in other places in frameworks.

REVIEW: 125772