Update docs for KCrash::Initialize.
Currently the platform plugin makes no such call so this is the only way to
ensure that the crash handler is always launched.

As discussed on the mailing list, simply linking to KCrash is not enough. Many
distributions build their packages with --as-needed which causes the link to be
broken as, without any library usage, it appears to be unused.

https://mail.kde.org/pipermail/kde-frameworks-devel/2015-September/026973.html
https://mail.kde.org/pipermail/kde-frameworks-devel/2015-November/028630.html

REVIEW: 126868