raise to core dump handlers when drkonqi is done
with the rise of useful core dump handlers such as systemd's coredump and
ubuntu's apport it is no longer useful to handle things exclusively in
drkonqi. it bypasses sysadmins as well as distros in debugging efforts,
putting the entire flow of information on us.

the new behavior instead checks if a core pattern executable is set and if
so re-raises the signal so the kernel jumps in and invokes the handler.

(this unfortunately means that the core will contain our kcrash frames,
but that seems hardly avoidable)

CHANGELOG: CMake option KCRASH_CORE_PATTERN_RAISE to forward to kernel
REVIEW: 128437