Disable kcrash when running under rr
When we are recording a process with mozilla's rr, then we do not
want to enable the kcrash handler. Otherwise, the app stays alive
after crashing until the user cancels kcrash, which is cumbersome
and unintuitive.

Better detect that we are running under rr through the documented
RUNNING_UNDER_RR env var and then disable kcrash. This way, the
app aborts directly and the user can inspect the crash via rr.