KCrash: remove debug output which breaks unittests from using ~/.qttest/config for categorized logging
Summary:
kdebugsettings --test-mode allows to tune the categorized logging for unittests,
but if a very early qCDebug happens, QLoggingRegistry::initializeRules is called
before the unittest gets a chance to call QStandardPaths::setTestModeEnabled(true).

This is obviously a bit fragile overall; should all unittests do that in a constructor function? A bit ugly...

Reviewers: kossebau, davidedmundson

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D29361