Fix KCrash usage inside a systemd unit
Systemd will send a SIGTERM to unit that are failing (such as
services with Type=dbus that drop off the bus), however if the process
is crashing and built with KCrash, the SIGTERM will end the process
cleanly while it is waiting for drkonqi to finish.

This patch makes the process crash with the original signal in this
case, so that systemd sees the service as crashed and can save the crash
dump, and restart if needed.