Don't make Qt5X11Extras a hard dep without X
77df85ed29e8304164f278a11de01b53d5c49dc6 duplicated the find_package
call from src/CMakeLists.txt. But that one was behind
"if (${X11_FOUND})" with X11 not being required, while the newly added
one is behind "if (UNIX ...)", turning Qt5X11Extras into a hard
dependency on Unices.
This wasn't the case before, isn't really needed and might be nicer in
the future with wayland becoming more prominent.