let some of the env vars disable all automatic init
when crash handling as a whole should be disabled by default we really
should not override that based on the metadata condition. rather, when
shouldHandleCrash() the crash handler never ought to get set up and
since metadata writing requires the handler by extension that means the
metadata ought not get written either

this is ultimately the similar to s_launchDrKonqi: the app opts out.
while shouldHandleCrash means: the env opts out.

in either case we shouldn't take any steps towards automatic
initialization