Add CMake option to build without X11
Currently using X11 libs is optional depending on whether the needed
libs are found.

This is quite brittle since it can lead to people accidentally not
having the right -dev packages installed and then getting surprised when
functionality is not there.

Since there are legitimate reasons to not build with X11 support add an
explicit option to disable it.