Fix @Qt5X11Extras_FOUND@ condition in cmake config file
Can't use 'TARGET Qt5::X11Extras' because it my be undefined at the time
this file is used; nor can we use @Qt5X11Extras_FOUND@ because it'll be
undefined on non-X11 or when building with Qt6, so the condition would
expand to an empty string, which isn't valid CMake code.

Instead use 'TARGET Qt5/6::Core' which will have been defined by the
find_dependency call a couple of line above.

NO_CHANGELOG