metadatainiwriter: move to metadata file naming v2
v2 is inspired by the coredumpd naming style which includes the pid, the
exe name, the bootid, but not the time of crash. this should provide us
with plenty of entropy to persistently store metadata on disk without
running risk of mistaking one crash for another. this in turn allows
drkonqi to report on crashes it hasn't yet seen without ever reporting
on a crash that it has already seen.

we exclude the time from the file name because getting it injected into
the name without malloc involves a whole lot of code. we should have
enough uniqueness without it as well