detect active exceptions and put them in the metadata
technically very straight forward except for a needed refactor of how
the INI writing works: instead of snprintfing into a preallocated string
we now write out the line on the fly. this allows us to make sure value
doesn't contain \n that would break the ini format. when a \n appears
we'll write it out as string rather than escape sequence.

CCBUG: 287458